+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.3.1-SNAPSHOT
 
 * [FEATURE] Version check at start
+* [FEATURE] Report prototype
+* [ENHANCEMENT] GUI improvements
 
 ## 2.3.0 / 2022-04-09